requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
anthropic>=0.20.0
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
pypdf>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
